% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readit.R
\name{guess_foreign}
\alias{guess_foreign}
\title{Guess File Type to Pass to \code{foreign} Readers}
\usage{
guess_foreign(.data)
}
\arguments{
\item{.data}{Data to guess/read}
}
\value{
A reader function, and its label
}
\description{
This function is a helper for \code{\link[=readit]{readit()}} to guess the type of file that can
be passed to an appropriate reader from \href{https://cran.r-project.org/package=foreign}{foreign}.
}
