% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readit.R
\name{guess_haven}
\alias{guess_haven}
\title{Guess File Type to Pass to Haven Readers}
\usage{
guess_haven(.data, ...)
}
\arguments{
\item{.data}{Data to guess/read}

\item{...}{Other arguments passed from parent function}
}
\value{
A reader function, and its label
}
\description{
This function is a helper for \code{\link[=readit]{readit()}} to guess the type of file that can
be passed to an appropriate reader from \href{http://haven.tidyverse.org/}{haven}.
}
