% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readit.R
\name{guess_txt}
\alias{guess_txt}
\title{Guess ".txt" Delimiter}
\usage{
guess_txt(.data)
}
\arguments{
\item{.data}{Data to guess/read}
}
\value{
A reader function, and its label
}
\description{
Since a \code{.txt} file is of ambiguous delimitation, this function is a helper
for \code{\link[=readit]{readit()}} to guess the type, and return the appropriate reader.
}
